+++ example/lib/app_model.dart
@@ -13,6 +13,8 @@ class AppModelImplementation extends AppModel {
 
   AppModelImplementation() {
     /// lets pretend we have to do some async initialization
+    /// using signalReady is not necessary in most cases
+    /// just use registerSingletonAsync instead
     Future.delayed(const Duration(seconds: 3))
         .then((_) => getIt.signalReady(this));
   }