+++ lib/get_it.dart
@@ -3,6 +3,7 @@
 library get_it;
 
 import 'dart:async';
+import 'dart:collection';
 
 import 'package:async/async.dart';
 import 'package:collection/collection.dart' show IterableExtension;
@@ -61,7 +62,7 @@ typedef FactoryFuncParamAsync<T, P1, P2> = Future<T> Function(
 );
 
 /// Data structure used to identify a dependency by type and instanceName
-class InitDependency extends Type {
+class InitDependency implements Type {
   final Type type;
   final String? instanceName;
 
@@ -157,9 +158,9 @@ abstract class GetIt {
   /// [type] if you want to get an instance by a Type object instead of a generic parameter.This should
   /// rarely be needed but can be useful if you have a runtime type and want to get an instance
   T get<T extends Object>({
-    String? instanceName,
     dynamic param1,
     dynamic param2,
+    String? instanceName,
     Type? type,
   });
 
@@ -359,14 +360,15 @@ abstract class GetIt {
   /// is registered inside GetIt
   bool isRegistered<T extends Object>({Object? instance, String? instanceName});
 
-  /// Clears all registered types. Handy when writing unit tests
+  /// Clears all registered types in the reverse order in which they were registered.
+  /// Handy when writing unit tests or when disposing services that depend on each other.
   /// If you provided dispose function when registering they will be called
   /// [dispose] if `false` it only resets without calling any dispose
   /// functions
   /// As dispose functions can be async, you should await this function.
   Future<void> reset({bool dispose = true});
 
-  /// Clears all registered types for the current scope
+  /// Clears all registered types for the current scope in the reverse order of registering them.
   /// If you provided dispose function when registering they will be called
   /// [dispose] if `false` it only resets without calling any dispose
   /// functions
@@ -423,7 +425,8 @@ abstract class GetIt {
   /// If no scope with [name] exists, nothing is popped and `false` is returned
   Future<bool> popScopesTill(String name, {bool inclusive = true});
 
-  /// Disposes all registered factories and singletons in the provided scope,
+  /// Disposes all registered factories and singletons in the provided scope
+  /// (in the reverse order in which they were registered),
   /// then destroys (drops) the scope. If the dropped scope was the last one,
   /// the previous scope becomes active again.
   /// if you provided dispose functions on registration, they will be called.